+++ erpnext/apps/rootedops_payroll/rootedops_payroll/rootedops_payroll/doctype/rootedops_nacha_profile/rootedops_nacha_profile.py
@@ -1,0 +1,8 @@
+from frappe.model.document import Document
+
+from rootedops_payroll.services.nacha_profile import validate_and_project_profile
+
+
+class RootedOpsNACHAProfile(Document):
+    def validate(self):
+        validate_and_project_profile(self)